test(select): show docs select tests
