+++ apps/website/src/routes/docs/headless/select/select.driver.ts
@@ -1,0 +1,25 @@
+import { Locator, Page } from '@playwright/test';
+
+export type DriverLocator = Locator | Page;
+
+export function selectTestDriver<T extends DriverLocator>(locator: T) {
+  const getRoot = () => {
+    return locator.getByRole('combobox');
+  };
+
+  return {
+    ...locator,
+    locator,
+    getRoot,
+    getListbox() {
+      return getRoot().getByRole('listbox');
+    },
+    getTrigger() {
+      return getRoot().getByRole('button');
+    },
+    // get all options
+    getOptions() {
+      return getRoot().getByRole('option').all();
+    },
+  };
+}
