+++ apps/website/tsconfig.json
@@ -26,5 +26,5 @@
       "path": "./tsconfig.spec.json"
     }
   ],
-  "include": ["src", "e2e/select.spec.ts"]
+  "include": ["src", "src/routes/docs/headless/select/select.spec.ts"]
 }
