+++ package.json
@@ -15,6 +15,7 @@
     "changeset": "changeset",
     "changeset.version": "changeset version && pnpm install --no-frozen-lockfile && git add --all",
     "dev": "nx serve website",
+    "playwright": "nx e2e website",
     "format.fix": "pretty-quick --staged",
     "link.dist": "cd dist/packages/kit-headless && pnpm link --global",
     "lint": "nx affected:lint",
